requirement.txt
openai
xmltodict
requests
python-dotenv
bs4